backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
alembic==1.12.1
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
celery==5.3.4
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
tweepy==4.14.0
instaloader==4.10.3
praw==7.7.1
pytesseract==0.3.10
python-telegram-bot==20.7
pycryptodome==3.19.0
cryptography==41.0.7
python-dateutil==2.8.2
pytz==2023.3
structlog==23.2.0
prometheus-client==0.19.0
selenium==4.15.2
webdriver-manager==4.0.1
imagehash==4.3.1
requests-html==0.10.0
scrapy==2.11.0
opencv-python-headless==4.8.1.78
requests==2.31.0
tqdm==4.66.1
google-cloud-vision==3.4.5
# face-recognition==1.3.0 # Removed to make optional
# dlib==19.24.2 # Removed to make optional

